refactor(ledger): collocate the balance route registrars :hammer:
Same move as the asset resource: the Huma terminal registrar came out of
balance_handler.go and the Fiber guard-chain registrars came out of routes.go,
and both now live in balance_routes.go.

The midaz pkg/net/http import takes the pkgHTTP alias so the file can also
import stdlib net/http for the Huma operation constants.

No route, authz tuple, operation ID or path changed.
