+++ components/ledger/internal/adapters/http/in/balance_handler.go
@@ -427,125 +427,3 @@ func (handler *BalanceHandler) GetAccountBalancesAtTimestamp(ctx context.Context
 
 	return &GetAccountBalanceHistoryResponse{Status: http.StatusOK, Body: history}, nil
 }
-
-// RegisterBalanceRoutes registers the ten balance operations on the
-// shared Huma API. The auth
-// (auth.Authorize("midaz","balances",verb)) + tenant + ParseUUIDPathParameters
-// ("balance") chain for these routes is attached in the unified server (Fiber
-// level) BEFORE the Huma terminal, not here. Paths are GROUP-RELATIVE (the Huma API
-// is bound to a versioned Fiber group; the group's PrefixModifier writes the version
-// into each op's op.Path, not into a servers entry).
-//
-// opSuffix distinguishes the operation IDs one version group publishes from another's —
-// see routeOpSuffixV1. A straight v1/v2 mirror reuses the same handler methods and the
-// same input/output types, so only the operation IDs differ between the twins.
-func RegisterBalanceRoutes(api huma.API, h *BalanceHandler, opSuffix string) {
-	const (
-		orgLedger      = "/organizations/{organization_id}/ledgers/{ledger_id}"
-		balancesPath   = orgLedger + "/balances"
-		balanceIDPath  = balancesPath + "/{balance_id}"
-		balanceHistory = balanceIDPath + "/history"
-		acctBalances   = orgLedger + "/accounts/{account_id}/balances"
-		acctHistory    = acctBalances + "/history"
-		aliasBalances  = orgLedger + "/accounts/alias/{alias}/balances"
-		codeBalances   = orgLedger + "/accounts/external/{code}/balances"
-		tag            = "Balances"
-	)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getAllBalances" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        balancesPath,
-		Summary:     "Get all balances",
-		Tags:        []string{tag},
-		Security:    secBalanceBearer,
-	}, h.GetAllBalances)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getBalanceByID" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        balanceIDPath,
-		Summary:     "Get Balance by id",
-		Tags:        []string{tag},
-		Security:    secBalanceBearer,
-	}, h.GetBalanceByID)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getBalanceAtTimestamp" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        balanceHistory,
-		Summary:     "Get Balance history at date",
-		Tags:        []string{tag},
-		Security:    secBalanceBearer,
-	}, h.GetBalanceAtTimestamp)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getAllBalancesByAccountID" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        acctBalances,
-		Summary:     "Get all balances by account id",
-		Tags:        []string{tag},
-		Security:    secBalanceBearer,
-	}, h.GetAllBalancesByAccountID)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getAccountBalancesAtTimestamp" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        acctHistory,
-		Summary:     "Get Account Balances history at date",
-		Tags:        []string{tag},
-		Security:    secBalanceBearer,
-	}, h.GetAccountBalancesAtTimestamp)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getBalancesByAlias" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        aliasBalances,
-		Summary:     "Get Balances using Alias",
-		Tags:        []string{tag},
-		Security:    secBalanceBearer,
-	}, h.GetBalancesByAlias)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getBalancesExternalByCode" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        codeBalances,
-		Summary:     "Get External balances using code",
-		Tags:        []string{tag},
-		Security:    secBalanceBearer,
-	}, h.GetBalancesExternalByCode)
-
-	huma.Register(api, huma.Operation{
-		OperationID:      "updateBalance" + opSuffix,
-		Method:           http.MethodPatch,
-		Path:             balanceIDPath,
-		Summary:          "Update Balance",
-		Tags:             []string{tag},
-		Security:         secBalanceBearer,
-		SkipValidateBody: true, // body validated imperatively — see file header.
-	}, h.UpdateBalance)
-	attachTypedRequestBody[mmodel.UpdateBalance](api, "updateBalance"+opSuffix)
-
-	huma.Register(api, huma.Operation{
-		OperationID:      "createAdditionalBalance" + opSuffix,
-		Method:           http.MethodPost,
-		Path:             acctBalances,
-		Summary:          "Create Additional Balance",
-		Tags:             []string{tag},
-		Security:         secBalanceBearer,
-		SkipValidateBody: true, // body validated imperatively — see file header.
-		DefaultStatus:    http.StatusCreated,
-	}, h.CreateAdditionalBalance)
-	attachTypedRequestBody[mmodel.CreateAdditionalBalance](api, "createAdditionalBalance"+opSuffix)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "deleteBalance" + opSuffix,
-		Method:      http.MethodDelete,
-		Path:        balanceIDPath,
-		Summary:     "Delete Balance by account",
-		Tags:        []string{tag},
-		Security:    secBalanceBearer,
-		// DefaultStatus 204 + an Out struct with no Body field => bodiless 204.
-		DefaultStatus: http.StatusNoContent,
-	}, h.DeleteBalanceByID)
-}
